+++ apps/astarte_housekeeping/lib/astarte_housekeeping/engine.ex
@@ -27,6 +27,7 @@ defmodule Astarte.Housekeeping.Engine do
         public_key_pem,
         replication_factor,
         device_registration_limit,
+        max_retention,
         opts \\ []
       ) do
     _ =
@@ -35,7 +36,8 @@ defmodule Astarte.Housekeeping.Engine do
         tag: "create_realm",
         realm: realm,
         replication_factor: replication_factor,
-        device_registration_limit: device_registration_limit
+        device_registration_limit: device_registration_limit,
+        datastream_maximum_storage_retention: max_retention
       )
 
     Queries.create_realm(
@@ -43,6 +45,7 @@ defmodule Astarte.Housekeeping.Engine do
       public_key_pem,
       replication_factor,
       device_registration_limit,
+      max_retention,
       opts
     )
   end
@@ -99,14 +102,15 @@ defmodule Astarte.Housekeeping.Engine do
 
   @doc """
   Updates a given realm using the provided values map. Fails if realm does not exists or values are invalid.
-  At the moment, updates to jwt_public_key_pem and device_registration_limit are supported.
+  At the moment, updates to jwt_public_key_pem, device_registration_limit and datastream_maximum_storage_retention are supported.
   Returns a tuple:
   - either {:ok, updated_realm} where updated_realm is a map describing the realm
   - or {:error, reason} where reason is an atom describing the error.
   """
   @spec update_realm(String.t(), %{
           :jwt_public_key_pem => String.t(),
-          :device_registration_limit => any(),
+          :device_registration_limit => non_neg_integer() | :remove_limit | nil,
+          :datastream_maximum_storage_retention => non_neg_integer() | nil,
           optional(any()) => any()
         }) ::
           {:ok, map()}
@@ -130,12 +134,14 @@ defmodule Astarte.Housekeeping.Engine do
   defp do_update_realm(realm_name, update_attrs) do
     %{
       jwt_public_key_pem: new_jwt_public_key_pem,
-      device_registration_limit: new_limit
+      device_registration_limit: new_limit,
+      datastream_maximum_storage_retention: new_retention
     } = update_attrs
 
     with realm when is_map(realm) <- Queries.get_realm(realm_name),
          :ok <- update_jwt_public_key_pem(realm_name, new_jwt_public_key_pem),
-         :ok <- update_device_registration_limit(realm_name, new_limit) do
+         :ok <- update_device_registration_limit(realm_name, new_limit),
+         :ok <- update_datastream_maximum_storage_retention(realm_name, new_retention) do
       updated_realm = merge_realm_status(realm, update_attrs)
 
       _ = Logger.info("Successful update of realm #{realm_name}", tag: "realm_update_success")
@@ -176,6 +182,31 @@ defmodule Astarte.Housekeeping.Engine do
     :ok
   end
 
+  defp update_datastream_maximum_storage_retention(realm_name, new_retention)
+       when is_integer(new_retention) and new_retention != 0 do
+    _ =
+      Logger.info("Updating datastream maximum storage retention",
+        tag: "update_datastream_maximum_storage_retention"
+      )
+
+    set_datastream_maximum_storage_retention(realm_name, new_retention)
+  end
+
+  # ScyllaDB considers TTL=0 as unset, see
+  # https://opensource.docs.scylladb.com/stable/cql/time-to-live.html#notes
+  defp update_datastream_maximum_storage_retention(realm_name, 0) do
+    _ =
+      Logger.info("Removing datastream maximum storage retention",
+        tag: "remove_datastream_maximum_storage_retention"
+      )
+
+    delete_datastream_maximum_storage_retention(realm_name)
+  end
+
+  defp update_datastream_maximum_storage_retention(_realm_name, nil) do
+    :ok
+  end
+
   defp delete_device_registration_limit(realm_name) do
     case Queries.delete_device_registration_limit(realm_name) do
       {:ok, %Xandra.Void{}} ->
@@ -189,6 +220,22 @@ defmodule Astarte.Housekeeping.Engine do
           )
 
         {:error, :delete_device_registration_limit_fail}
+    end
+  end
+
+  defp delete_datastream_maximum_storage_retention(realm_name) do
+    case Queries.delete_datastream_maximum_storage_retention(realm_name) do
+      {:ok, %Xandra.Void{}} ->
+        :ok
+
+      {:error, reason} ->
+        _ =
+          Logger.warning(
+            "Cannot delete datastream maximum storage retention, error #{inspect(reason)}",
+            tag: "delete_datastream_maximum_storage_retention"
+          )
+
+        {:error, :delete_datastream_maximum_storage_retention_fail}
     end
   end
 
@@ -208,6 +255,23 @@ defmodule Astarte.Housekeeping.Engine do
     end
   end
 
+  defp set_datastream_maximum_storage_retention(realm_name, new_retention)
+       when is_integer(new_retention) do
+    case Queries.set_datastream_maximum_storage_retention(realm_name, new_retention) do
+      {:ok, %Xandra.Void{}} ->
+        :ok
+
+      {:error, reason} ->
+        _ =
+          Logger.warning(
+            "Cannot set datastream maximum storage retention, error #{inspect(reason)}",
+            tag: "set_datastream_maximum_storage_retention"
+          )
+
+        {:error, :set_datastream_maximum_storage_retention_fail}
+    end
+  end
+
   defp is_realm_update_valid?(update_attrs) do
     # TODO from ScyllaDB >= 5.3, replication can be altered
     update_valid? =
@@ -224,6 +288,8 @@ defmodule Astarte.Housekeeping.Engine do
     update_valid?
   end
 
+  # If device_registration_limit is nil, it means it was not present in the update
+  # payload. Act accordingly.
   defp merge_realm_status(old_realm, %{device_registration_limit: nil} = update_attrs) do
     attrs = Map.delete(update_attrs, :device_registration_limit)
     merge_realm_status(old_realm, attrs)
@@ -232,6 +298,23 @@ defmodule Astarte.Housekeeping.Engine do
   defp merge_realm_status(old_realm, %{device_registration_limit: :remove_limit} = update_attrs) do
     attrs = Map.delete(update_attrs, :device_registration_limit)
     merge_realm_status(%{old_realm | device_registration_limit: nil}, attrs)
+  end
+
+  # If datastream_maximum_storage_retention is nil, it means it was not present in the update
+  # payload. Act accordingly.
+  defp merge_realm_status(old_realm, %{datastream_maximum_storage_retention: nil} = update_attrs) do
+    attrs = Map.delete(update_attrs, :datastream_maximum_storage_retention)
+    merge_realm_status(old_realm, attrs)
+  end
+
+  # Differently from device_registration_limit, datastream_maximum_storage_retention
+  # uses 0 to unset the value. This follows the convention used by ScyllaDB/Cassandra.
+  defp merge_realm_status(
+         old_realm,
+         %{datastream_maximum_storage_retention: 0} = update_attrs
+       ) do
+    attrs = Map.delete(update_attrs, :datastream_maximum_storage_retention)
+    merge_realm_status(%{old_realm | datastream_maximum_storage_retention: nil}, attrs)
   end
 
   defp merge_realm_status(old_realm, update_attrs) do