+++ apps/astarte_housekeeping_api/lib/astarte_housekeeping_api/rpc/housekeeping.ex
@@ -49,7 +49,8 @@ defmodule Astarte.Housekeeping.API.RPC.Housekeeping do
           jwt_public_key_pem: pem,
           replication_class: "SimpleStrategy",
           replication_factor: replication_factor,
-          device_registration_limit: device_registration_limit
+          device_registration_limit: device_registration_limit,
+          datastream_maximum_storage_retention: datastream_maximum_storage_retention
         },
         opts
       ) do
@@ -59,7 +60,8 @@ defmodule Astarte.Housekeeping.API.RPC.Housekeeping do
       jwt_public_key_pem: pem,
       replication_class: :SIMPLE_STRATEGY,
       replication_factor: replication_factor,
-      device_registration_limit: device_registration_limit
+      device_registration_limit: device_registration_limit,
+      datastream_maximum_storage_retention: datastream_maximum_storage_retention
     }
     |> encode_call(:create_realm)
     |> @rpc_client.rpc_call(@destination, Config.rpc_timeout!())
@@ -73,7 +75,8 @@ defmodule Astarte.Housekeeping.API.RPC.Housekeeping do
           jwt_public_key_pem: pem,
           replication_class: "NetworkTopologyStrategy",
           datacenter_replication_factors: replication_factors_map,
-          device_registration_limit: device_registration_limit
+          device_registration_limit: device_registration_limit,
+          datastream_maximum_storage_retention: datastream_maximum_storage_retention
         },
         opts
       ) do
@@ -83,7 +86,8 @@ defmodule Astarte.Housekeeping.API.RPC.Housekeeping do
       jwt_public_key_pem: pem,
       replication_class: :NETWORK_TOPOLOGY_STRATEGY,
       datacenter_replication_factors: Enum.to_list(replication_factors_map),
-      device_registration_limit: device_registration_limit
+      device_registration_limit: device_registration_limit,
+      datastream_maximum_storage_retention: datastream_maximum_storage_retention
     }
     |> encode_call(:create_realm)
     |> @rpc_client.rpc_call(@destination)
@@ -97,7 +101,8 @@ defmodule Astarte.Housekeeping.API.RPC.Housekeeping do
         replication_class: replication_class,
         replication_factor: replication_factor,
         datacenter_replication_factors: replication_factors_map,
-        device_registration_limit: limit
+        device_registration_limit: limit,
+        datastream_maximum_storage_retention: max_retention
       }) do
     %UpdateRealm{
       realm: realm_name,
@@ -105,7 +110,9 @@ defmodule Astarte.Housekeeping.API.RPC.Housekeeping do
       replication_class: replication_class_to_atom(replication_class),
       replication_factor: replication_factor,
       datacenter_replication_factors: replication_factors_map,
-      device_registration_limit: device_registration_limit_to_proto(limit)
+      device_registration_limit: device_registration_limit_to_proto(limit),
+      datastream_maximum_storage_retention:
+        datastream_maximum_storage_retention_to_proto(max_retention)
     }
     |> encode_call(:update_realm)
     |> @rpc_client.rpc_call(@destination, Config.rpc_timeout!())
@@ -192,7 +199,8 @@ defmodule Astarte.Housekeeping.API.RPC.Housekeeping do
             jwt_public_key_pem: pem,
             replication_class: :SIMPLE_STRATEGY,
             replication_factor: replication_factor,
-            device_registration_limit: device_registration_limit
+            device_registration_limit: device_registration_limit,
+            datastream_maximum_storage_retention: max_retention
           }}
        ) do
     {:ok,
@@ -201,7 +209,9 @@ defmodule Astarte.Housekeeping.API.RPC.Housekeeping do
        jwt_public_key_pem: pem,
        replication_class: "SimpleStrategy",
        replication_factor: replication_factor,
-       device_registration_limit: device_registration_limit
+       device_registration_limit: device_registration_limit,
+       datastream_maximum_storage_retention:
+         datastream_maximum_storage_retention_from_proto(max_retention)
      }}
   end
 
@@ -212,7 +222,8 @@ defmodule Astarte.Housekeeping.API.RPC.Housekeeping do
             jwt_public_key_pem: pem,
             replication_class: :NETWORK_TOPOLOGY_STRATEGY,
             datacenter_replication_factors: datacenter_replication_factors,
-            device_registration_limit: device_registration_limit
+            device_registration_limit: device_registration_limit,
+            datastream_maximum_storage_retention: max_retention
           }}
        ) do
     {:ok,
@@ -221,7 +232,9 @@ defmodule Astarte.Housekeeping.API.RPC.Housekeeping do
        jwt_public_key_pem: pem,
        replication_class: "NetworkTopologyStrategy",
        datacenter_replication_factors: Enum.into(datacenter_replication_factors, %{}),
-       device_registration_limit: device_registration_limit
+       device_registration_limit: device_registration_limit,
+       datastream_maximum_storage_retention:
+         datastream_maximum_storage_retention_from_proto(max_retention)
      }}
   end
 
@@ -239,6 +252,20 @@ defmodule Astarte.Housekeeping.API.RPC.Housekeeping do
          {:generic_error_reply, %GenericErrorReply{error_name: "connected_devices_present"}}
        ) do
     {:error, :connected_devices_present}
+  end
+
+  defp extract_reply(
+         {:generic_error_reply,
+          %GenericErrorReply{error_name: "delete_datastream_maximum_storage_retention_fail"}}
+       ) do
+    {:error, :delete_datastream_maximum_storage_retention_fail}
+  end
+
+  defp extract_reply(
+         {:generic_error_reply,
+          %GenericErrorReply{error_name: "set_datastream_maximum_storage_retention_fail"}}
+       ) do
+    {:error, :set_datastream_maximum_storage_retention_fail}
   end
 
   defp extract_reply({:generic_error_reply, error_struct = %GenericErrorReply{}}) do
@@ -271,10 +298,16 @@ defmodule Astarte.Housekeeping.API.RPC.Housekeeping do
   defp replication_class_to_atom("SimpleStrategy"), do: :SIMPLE_STRATEGY
   defp replication_class_to_atom(nil), do: nil
 
-  defp device_registration_limit_to_proto(:remove_limit), do: {:remove_limit, %RemoveLimit{}}
+  defp device_registration_limit_to_proto(:unset), do: {:remove_limit, %RemoveLimit{}}
   defp device_registration_limit_to_proto(nil), do: nil
 
   defp device_registration_limit_to_proto(n) when is_integer(n) do
     {:set_limit, %SetLimit{value: n}}
   end
+
+  defp datastream_maximum_storage_retention_to_proto(:unset), do: 0
+  defp datastream_maximum_storage_retention_to_proto(nil), do: nil
+  defp datastream_maximum_storage_retention_to_proto(n) when is_integer(n), do: n
+  defp datastream_maximum_storage_retention_from_proto(0), do: nil
+  defp datastream_maximum_storage_retention_from_proto(n) when is_integer(n), do: n
 end