+++ apps/astarte_housekeeping_api/lib/astarte_housekeeping_api_web/views/realm_view.ex
@@ -40,7 +40,8 @@ defmodule Astarte.Housekeeping.APIWeb.RealmView do
         jwt_public_key_pem: realm.jwt_public_key_pem,
         replication_class: "SimpleStrategy",
         replication_factor: realm.replication_factor,
-        device_registration_limit: realm.device_registration_limit
+        device_registration_limit: realm.device_registration_limit,
+        datastream_maximum_storage_retention: realm.datastream_maximum_storage_retention
       }
     }
   end
@@ -52,7 +53,8 @@ defmodule Astarte.Housekeeping.APIWeb.RealmView do
         jwt_public_key_pem: realm.jwt_public_key_pem,
         replication_class: "NetworkTopologyStrategy",
         datacenter_replication_factors: realm.datacenter_replication_factors,
-        device_registration_limit: realm.device_registration_limit
+        device_registration_limit: realm.device_registration_limit,
+        datastream_maximum_storage_retention: realm.datastream_maximum_storage_retention
       }
     }
   end